+++ src/Channel.ts
@@ -5,9 +5,9 @@ interface BroadcastOptions {
 	/**
 	 * Filter sessions that should receive the event.
 	 *
-	 * Called with each session and should return a truthy value to allow the event to be sent, otherwise return a falsy value to prevent the session from receiving the event.
+	 * Called with each session and should return `true` to allow the event to be sent and otherwise return `false` to prevent the session from receiving the event.
 	 */
-	filter?: (session: Session) => unknown;
+	filter?: (session: Session) => boolean;
 }
 
 interface Events extends EventMap {